CellExtra now holds a Vec of graphics to draw
This allows multiple graphics to be anchored to the same cell. Without
this, graphics can still overlap as long as they are not all started
in the same cell.

SmallVec<GraphicCell> would probably be a better choice than the
Option<Vec<GraphicCell>> that I’ve used, but this is a smaller change.

It would also be nice not to clone the graphics when we display them;
perhaps we should mix in an Arc or something.