implement a way to deliberately remove sixel images
The original specification assumes that printing a sixel modifies the
pixels in a framebuffer destructively. Printing another sixel over top
of an existing one would both destructively overwrite pixels (barring
transparency in the second image).

Alacritty on the other hand keeps the whole stack of images around and
sends them off to the graphics card to be composited together. The
second image can be drawn over the first, and this will be visually
correct. However, both images are still sent to the video card. This
means that clearing an image from the screen means overwriting it,
hopefully with the correct background color, and accepting the
performance penalty of keeping both images around forever.

Alacritty’s implementation first held a single image per character
cell, so printing a sixel image with no pixels would efficiently
delete the existing image. With the change to hold a list of images so
that transparency can be used in a manner similar to the original DEC
implementation, this no longer works.

With this patch Alacritty will look at the macro parameter of the
sixel control string. This is normally always zero, but if 100 is used
instead then we will first delete all images anchored to the current
cursor position. The new graphic will still potentially overlap with
other graphics that have not been deleted.

Note that we still cannot overwrite a sixel with text; that will
require more complicated work.